Separate out gRPC utility functions
